put pagination to utils for reuse
